Added weapon and skills

+++ Services/CharacterService/CharacterService.cs
@@ -1,11 +1,14 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using dotnet_rpg.Controllers.Dtos.Character;
 using dotnet_rpg.Data;
+using dotnet_rpg.Dtos.Character;
 using dotnet_rpg.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace dotnet_rpg.Services.CharacterService
@@ -15,20 +18,28 @@ namespace dotnet_rpg.Services.CharacterService
         private readonly IMapper mapper;
 
         private readonly DataContext context;
+        private readonly IHttpContextAccessor httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
+            this.httpContextAccessor = httpContextAccessor;
             this.context = context;
             this.mapper = mapper;
         }
+        //Dobijanje logovanog usera pomocu tokena
+        private int GetUserId() => int.Parse(httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = mapper.Map<Character>(newCharacter);
+            character.User = await context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
             context.Characters.Add(character);
             await context.SaveChangesAsync();
-            serviceResponse.Data = await context.Characters.Select(c => mapper.Map<GetCharacterDto>(c)).ToListAsync();
+            serviceResponse.Data = await context.Characters
+            .Where(c => c.User.Id == GetUserId())
+            .Select(c => mapper.Map<GetCharacterDto>(c)).ToListAsync();
             return serviceResponse;
         }
 
@@ -37,11 +48,20 @@ namespace dotnet_rpg.Services.CharacterService
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character = await context.Characters.FirstAsync(c => c.Id == id);
+                Character character = await context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if(character != null)
+                {            
                 context.Characters.Remove(character);
                 await context.SaveChangesAsync();
-                serviceResponse.Data = context.Characters.Select(c => mapper.Map<GetCharacterDto>(c)).ToList();
-
+                serviceResponse.Data = context.Characters
+                .Where(c => c.User.Id == GetUserId())
+                .Select(c => mapper.Map<GetCharacterDto>(c)).ToList();
+                
+                }else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not fount.";
+                }
             }
             catch (Exception ex)
             {
@@ -51,11 +71,14 @@ namespace dotnet_rpg.Services.CharacterService
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             //Dobijanje karaktera koji su povezani za jednog korisnika preko id - a
-            var dbCharacters = await context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+            var dbCharacters = await context.Characters
+                .Include(c => c.Weapon)
+                .Include(c => c.Skills)
+                .Where(c => c.User.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = dbCharacters.Select(c => mapper.Map<GetCharacterDto>(c)).ToList();
             return serviceResponse;
         }
@@ -63,7 +86,10 @@ namespace dotnet_rpg.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await context.Characters
+                .Include(c => c.Weapon)
+                .Include(c => c.Skills)
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
@@ -73,19 +99,26 @@ namespace dotnet_rpg.Services.CharacterService
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
             try
             {
-                Character character = await context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
-
-                character.Name = updatedCharacter.Name;
-                character.HitPoints = updatedCharacter.HitPoints;
-                character.Strength = updatedCharacter.Strength;
-                character.Defense = updatedCharacter.Defense;
-                character.Intelligence = updatedCharacter.Intelligence;
-                character.Class = updatedCharacter.Class;
+                Character character = await context.Characters
+                .Include(c => c.User)
+                .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+                if(character.User.Id == GetUserId())
+                {   character.Name = updatedCharacter.Name;
+                    character.HitPoints = updatedCharacter.HitPoints;
+                    character.Strength = updatedCharacter.Strength;
+                    character.Defense = updatedCharacter.Defense;
+                    character.Intelligence = updatedCharacter.Intelligence;
+                    character.Class = updatedCharacter.Class;
 
                 await context.SaveChangesAsync();
 
                 serviceResponse.Data = mapper.Map<GetCharacterDto>(character);
-
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found.";
+                }
             }
             catch (Exception ex)
             {
@@ -93,6 +126,44 @@ namespace dotnet_rpg.Services.CharacterService
                 serviceResponse.Message = ex.Message;
             }
             return serviceResponse;
+        }
+
+        public async Task<ServiceResponse<GetCharacterDto>> AddCharacterSkill(AddCharacterSkillDto newCharacterSkill)
+        {
+            var response = new ServiceResponse<GetCharacterDto>();
+
+            try
+            {
+                var character = await context.Characters
+                    .Include(c => c.Weapon)
+                    .Include(c => c.Skills)
+                    .FirstOrDefaultAsync(c => c.Id == newCharacterSkill.CharacterId && c.User.Id == GetUserId());
+
+                    if(character == null)
+                    {
+                        response.Success = false;
+                        response.Message = "Character not found.";
+                        return response;
+                    }
+
+                    var skill = await context.Skills.FirstOrDefaultAsync(s => s.Id == newCharacterSkill.SkillId);
+                    if(skill == null)
+                    {
+                        response.Success = false;
+                        response.Message = "Skill not found.";
+                        return response;
+                    }
+                    character.Skills.Add(skill);
+                    await context.SaveChangesAsync();
+
+                    response.Data = mapper.Map<GetCharacterDto>(character);
+            }
+            catch(Exception ex)
+            {
+                response.Success = false;
+                response.Message = ex.Message;
+            }
+            return response;
         }
     }
 }
